requirements.txt
yt-dlp
yadisk
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-slugify
python-dotenv
pydantic>=2.0.0
pydantic-settings>=1.0.0
PySide6
pytest
pytest-qt
requests==2.32.3
httpx==0.27.0
tenacity==8.5.0
yadisk_async==1.4.5